+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add new tool to auto generate an OpenAPI v3 compatible specification with all AEM configuration nodes
 
 ## 3.0.0 - 2019-05-14
 ### Added